move ResizeSensor to popover2 package:
