+++ packages/popover2/src/index.md
@@ -31,3 +31,4 @@ Import CSS with a bundler like webpack:
 @page popover2
 @page tooltip2
 @page context-menu2
+@page resize-sensor2
