+++ packages/popover2/src/index.ts
@@ -36,4 +36,5 @@ export {
     StrictModifierNames,
 } from "./popover2SharedProps";
 export { IPopover2Props, Popover2Props, Popover2, Popover2InteractionKind } from "./popover2";
+export { ResizeSensor2, ResizeSensor2Props } from "./resizeSensor2";
 export { ITooltip2Props, Tooltip2Props, Tooltip2 } from "./tooltip2";
