+++ packages/popover2/test/index.ts
@@ -17,5 +17,6 @@ import "@blueprintjs/test-commons/bootstrap";
 
 import "./contextMenu2Tests";
 import "./popover2Tests";
+import "./resizeSensor2Tests";
 import "./tooltip2Tests";
 import "./utilsTests";
